split tests and fix formatting
